wip: refactored code
